v2.3
svn path=/branches/V_2_3/; revision=290

+++ test/emailrelay_test_client.cpp
@@ -181,7 +181,7 @@ void Test::sendMessage( int lines , int length , bool last )
 	send( "DATA\r\n" ) ;
 	waitline() ;
 
-	std::vector<char> buffer( static_cast<unsigned>(length+2) ) ;
+	std::vector<char> buffer( static_cast<std::size_t>(length)+2U ) ;
 	std::fill( buffer.begin() , buffer.end() , 't' ) ;
 	buffer[buffer.size()-1U] = '\n' ;
 	buffer[buffer.size()-2U] = '\r' ;
